Code Quality: Moved service interfaces to Contracts & Removed JumpListHelper (#15148)
