+++ src/Files.App/Data/Contracts/IPreviewPopupService.cs
@@ -1,7 +1,7 @@
 ﻿// Copyright (c) 2024 Files Community
 // Licensed under the MIT License. See the LICENSE.
 
-namespace Files.App.Services
+namespace Files.App.Data.Contracts
 {
 	public interface IPreviewPopupService : INotifyPropertyChanged
 	{
