+++ src/Files.App/Data/Contracts/IStartMenuService.cs
@@ -1,6 +1,6 @@
 ﻿using Files.Core.Storage;
 
-namespace Files.App.Services
+namespace Files.App.Data.Contracts
 {
 	/// <summary>
 	/// A service that manages actions associated with system Start Menu.
