+++ src/Files.App/Data/Contracts/IWindowsJumpListService.cs
@@ -1,10 +1,12 @@
 ﻿// Copyright (c) 2024 Files Community
 // Licensed under the MIT License. See the LICENSE.
 
-namespace Files.App.Services
+namespace Files.App.Data.Contracts
 {
-	public interface IJumpListService
+	public interface IWindowsJumpListService
 	{
+		Task InitializeAsync();
+
 		Task AddFolderAsync(string path);
 
 		Task RefreshPinnedFoldersAsync();
